Add Global Parent to IOPub Messages
Before this commit, IOPub messages had no parent headers, and that
caused messages to be dropped by some frontends (E.G. notebook).
Because of that, using the notebook as a frontend was broken.

This commit implements a parent header for all IOPub messages.
The way it achieves that, is by setting a 'global' parent header that
is switched by the kernel according to incoming requests.
This is a hacky way to fix the problem, as multiple requests can run at
the same time, so we should find a better solution later.
All in all, this seems to work..