Implement Inspection Proxy Via Execution Requests
According to the jupyter docs, frontends can send inspection requests.
Nevertheless, I could not get trigger a frontend to send such request.
The python kernel returns documentation on execution request that end
with '?'. So I improved the execution request handler to redirect the
request to the inspection handler on '?', thus giving it a chance to run